+++ compiler/test/in/string_literal_formatting.js
@@ -1,0 +1,1 @@
+print('a special string with " and \\');